requirements.txt
# 核心依赖
streamlit>=1.28.0
# langchain>=0.1.0p
langchain==0.3.24
langchain-community>=0.0.10
openai>=1.3.0
chromadb>=0.4.0
tiktoken>=0.5.0

# 数据模型和验证
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 配置文件
pyyaml>=6.0
python-dotenv>=1.0.0

# HTTP请求
requests>=2.31.0
aiohttp>=3.9.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 文本处理
jieba>=0.42.1  # 中文分词
zhon>=1.1.5    # 中文标点

# 异步和并发
asyncio>=3.4.3
concurrent-log-handler>=0.9.20

# 工具函数
tqdm>=4.65.0
colorama>=0.4.6
python-dateutil>=2.8.2

# 开发工具（可选）
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0